refactor: BaseDTO 이용하여 DTO 중복 제거
